[APPC-904] 1-Change convertion response. 2-Add minor stuctural update. 3- Addresizable edit text.

+++ app/src/main/java/com/asfoundation/wallet/ui/iab/FiatValue.kt
@@ -1,5 +1,6 @@
 package com.asfoundation.wallet.ui.iab
 
 import java.io.Serializable
+import java.math.BigDecimal
 
-class FiatValue(val amount: Double, val currency: String, val symbol: String = "") : Serializable
+class FiatValue(val amount: BigDecimal, val currency: String, val symbol: String = "") : Serializable
